% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level_set_functions.R
\name{cate_lvl_set}
\alias{cate_lvl_set}
\title{cate_lvl_set}
\usage{
cate_lvl_set(
  theta,
  cate.obj,
  set_type = "upper",
  se = TRUE,
  B = 1000,
  alpha = 0.05
)
}
\arguments{
\item{theta}{the user-specified level}

\item{cate.obj}{a list of predictions of the cate at v0, fitted cate values,
residuals, design matrix at v0 and design matrix at oberved V.}

\item{set_type}{string indicating if upper level set, lower level set or level set
should be computed. Currently ignored.}

\item{se}{boolean for whether confidence sets should be returned.}

\item{B}{number of bootstrap replications, ignored if se = FALSE.}

\item{alpha}{1-confidence level, ignored if se = FALSE.}
}
\value{
a list containing the following components:
\item{level_set}{ the estimated level sets}
\item{chat.l}{ lower confidence set}
\item{chat.u}{ upper confidence set}
\item{cutoff}{ cutoff from multiplier bootstrap}
}
\description{
This function estimates the upper level set of the cate
}
