% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CD.SuperLearner.R
\name{CD.SuperLearner}
\alias{CD.SuperLearner}
\title{CD.SuperLearner}
\usage{
CD.SuperLearner(
  X,
  W,
  SL.library,
  n.folds = 10,
  n.bins = 2:floor(length(X)/50),
  verbose = FALSE,
  save.threshold = 0.001
)
}
\description{
CD.SuperLearner
}
