% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_learner.R
\name{dr_learner}
\alias{dr_learner}
\title{DR-Learner}
\usage{
dr_learner(x0, y, a, x, drl.x, mu1.x, mu0.x, pi.x, nsplits = 5)
}
\arguments{
\item{x0}{evaluation points, i.e. E(Y^1 - Y^0 | x0)}

\item{y}{vector of outcomes}

\item{a}{vector of treatments}

\item{x}{matrix of covariates}

\item{drl.x}{a function with arguments y, x, new.x computing the regression
of y on x and evaluating it at new.x. This is the model for the second-stage
regression.}

\item{mu1.x}{a function with arguments y, x, new.x computing the regression
of y on x among the treated and evaluating it at new.x.}

\item{mu0.x}{a function with arguments y, x, new.x computing the regression
of y on x among the un-treated and evaluating it at new.x.}

\item{pi.x}{a function with arguments a, x, new.x computing
the propensity score and evaluating it at new.x.}

\item{nsplits}{a number indicating the number of splits used to do cross-fitting}
}
\value{
a list containing the following components:
\item{est}{ estimate of the CATE at x0}
\item{fold_k_est}{ length(x0)xnsplits matrix with estimates of the CATE at x0 in each fold.}
}
\description{
This functions computes the DR-Learner approach to CATE estimation.
}
\examples{
drl.x <- function(y, x, new.x) {
dat <- cbind(data.frame(y = y), x)
predict(lm(y ~ ., data = dat), newdata = new.x)
}
mu1.x <- mu0.x <- drl.x
pi.x <- function(a, x, new.x) {
dat = cbind(data.frame(a = a), x)
predict(glm(a ~ ., data = dat, family = binomial()),
newdata = new.x, type = "response")
}
dr_learner(x0 = cbind(seq(0, 1, 0.1), seq(0, 1, 0.1)),
           y = rnorm(100), a = rbinom(100, 1, 0.5),
           x = matrix(runif(200), ncol = 2, nrow = 100),
           drl.x = drl.x, mu1.x = mu1.x, mu0.x = mu0.x, pi.x = pi.x,
           nsplits = 2)

}
\references{
Kennedy, EH. (2020). Optimal Doubly Robust Estimation of
Heterogeneous Causal Effects. \emph{arXiv preprint arXiv:2004.14497}.
}
