% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance_functions.R
\name{drl.basis.additive}
\alias{drl.basis.additive}
\title{drl.basis.additive
This function fits a low dimensional additive model}
\usage{
drl.basis.additive(y, x, new.x, kmin = 3, kmax = 10)
}
\arguments{
\item{y}{a numeric vector of outcomes}

\item{x}{a matrix or data frame of covariates' values}

\item{new.x}{a matrix or data frame of evaluation points}

\item{kmin}{minimum number of basis terms (for each covariate) to try in LOOCV step}

\item{kmax}{maximum number of basis terms (for each covariate) to try in LOOCV step.
A total of kmin*kmax model will be evaluated by LOOCV.}
}
\value{
All the fits as well as the best model
}
\description{
drl.basis.additive
This function fits a low dimensional additive model
}
