% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pseudo_outcome.R
\name{get.pseudo.y}
\alias{get.pseudo.y}
\title{unif_ctff_series}
\usage{
get.pseudo.y(
  y,
  a,
  x,
  v1,
  v2,
  mu0.x,
  mu1.x,
  pi.x,
  cond.dens,
  cate.w,
  nsplits = 1
)
}
\arguments{
\item{y}{outcome}

\item{a}{treatment}

\item{x}{confounders}

\item{v1}{effect modifiers 1}

\item{v2}{effect modifiers 2}

\item{cond.dens}{function for the conditional density of any of random variable
in v1 given all variables in v2}

\item{cate.w}{function for fitting E(tau(X) | v2)}

\item{nsplits}{number of splints}

\item{mu0x}{function for fitting mu0}

\item{mu1x}{function for fitting mu1}

\item{pix}{function for fitting propensity score}
}
\value{
a list containing the pseudo outcomes
}
\description{
This function calculates the pseudo-outcomes depending on the estimand
}
