% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance_functions.R
\name{get.smooth.fit.gam}
\alias{get.smooth.fit.gam}
\title{get.smooth.fit.gam
This function isolates the invdividual smooth components in low dimensional gam fit}
\usage{
get.smooth.fit.gam(fit, eval.pts, eff.modif.name, v)
}
\arguments{
\item{fit}{output from a low dimensional lm fit where the GAM is stored}

\item{eval.pts}{the evaluation points for the additive component of interst}

\item{eff.modif.name}{the name of the effect modifier}

\item{v}{the matrix of effect modifiers values
(the original data subsetted to the effect modifiers considered enetering the GAM)}
}
\value{
matrix of results: estimates and pointwise CIs.
}
\description{
get.smooth.fit.gam
This function isolates the invdividual smooth components in low dimensional gam fit
}
