% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp_r_learner.R
\name{lp_r_learner}
\alias{lp_r_learner}
\title{Lp-R-Learner}
\usage{
lp_r_learner(x0, y, a, x, mu.x, pi.x, basis, order_basis, kernel)
}
\arguments{
\item{x0}{evaluation points, i.e. E(Y^1 - Y^0 | x0)}

\item{y}{vector of outcomes}

\item{a}{vector of treatments}

\item{x}{matrix of covariates}

\item{mu.x}{a function with arguments y, x, new.x computing the regression
of y on x and evaluating it at new.x.}

\item{pi.x}{a function with arguments a, x, new.x computing
the propensity score and evaluating it at new.x.}

\item{basis}{a function with arguments x and j returning the j^th basis element
applied to x, e.g. x^j. It will be the building block to compute a tensor product basis.}

\item{order_basis}{the order of the basis}

\item{kernel}{a function with arguments x and x_0 returning K((x - x0) / h) / h}
}
\value{
a list containing the following components:
\item{est}{ estimate of the CATE at x0}
\item{fold_k_est}{ length(x0)xnsplits matrix with estimates of the CATE at x0 in each fold.}
}
\description{
This function computes the Lp-R-Learner approach to CATE estimation.
}
