% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debiased_inference_fns.R
\name{plot_debiased_curve}
\alias{plot_debiased_curve}
\title{plot_debiased_curve}
\usage{
plot_debiased_curve(
  pseudo,
  exposure,
  res.df,
  ci = TRUE,
  unif = TRUE,
  add.pseudo = TRUE
)
}
\arguments{
\item{res.df}{ADD}

\item{ci}{ADD}

\item{unif}{ADD}
}
\description{
plot_debiased_curve
}
