% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CD.SuperLearner.R
\name{predict.CD.SuperLearner}
\alias{predict.CD.SuperLearner}
\title{predict.CD.SuperLearner}
\usage{
\method{predict}{CD.SuperLearner}(fit, newdata, threshold = 0.001, verbose = FALSE)
}
\description{
predict.CD.SuperLearner
}
