% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cate.lvl.set.R
\name{unif_ctff_series}
\alias{unif_ctff_series}
\title{unif_ctff_series}
\usage{
unif_ctff_series(design.mat.v0, residuals, design.mat, B, alpha)
}
\arguments{
\item{design.mat.v0}{design matrix of evaluation points.}

\item{residuals}{observed residuals, i.e. pseudo.y - fitted.}

\item{B}{number of bootstrap replications}

\item{alpha}{1-confidence level, e.g. alpha = 0.05.}

\item{design.mat.v}{design matrix of observed v points.}
}
\value{
a list containing the following components:
\item{cutoff}{ the estimated quantile}
\item{sigma.hat}{ vector of estimated standard devition of sigma(v)}
\item{Omega.hat.sqrt} { Omega matrix }
}
\description{
This function calculates the quantile of sup_v|tauhat(v) - tau(v)| / sigma(v)
when tauhat(x) is based on a orthogonal series second stage regression.
}
\references{
Chernozhukov et al (2014).
}
