% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cate.R
\name{cate}
\alias{cate}
\title{CATE}
\usage{
cate(
  data_frame,
  learner,
  x_names,
  y_name,
  a_name,
  v_names,
  num_grid = 100,
  nsplits = 5,
  foldid = NULL,
  v0.long = NULL,
  univariate_reg = FALSE,
  partial_dependence = FALSE,
  partially_linear = FALSE,
  additive_approx = FALSE,
  variable_importance = FALSE,
  vimp_num_splits = 1,
  bw.stage2 = NULL,
  sample.split.cond.dens = FALSE,
  ...
)
}
\arguments{
\item{learner}{string specifying which learners to use, e.g. DR-Learner.}

\item{nsplits}{a number indicating the number of splits used to do
cross-fitting. Ignored if foldid is specified.}

\item{foldid}{id vector specifying which observation belongs to which fold.}

\item{v0}{evaluation points, i.e. E(Y^1 - Y^0 | V = v0)}

\item{y}{vector of outcomes}

\item{a}{binary vector of treatments}

\item{x}{matrix of confounders}

\item{v}{matrix of effect modifiers}
}
\value{
a list containing the following components:
\item{est}{ estimates of the CATE at v0}
\item{fold_est}{ list of length(v0)xnsplits matrices with estimates
of the CATE at v0 in each fold, one matrix for each learner.}
}
\description{
This function is a wrapper to estimate HTEs defined as E(Y^1 - Y^0 | V = v0)
}
\references{
Kennedy, EH. (2020). Optimal Doubly Robust Estimation of
Heterogeneous Causal Effects. \emph{arXiv preprint arXiv:2004.14497}.
}
