% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debiased_inference_fns.R
\name{debiased_inference}
\alias{debiased_inference}
\title{debiased_inference}
\usage{
debiased_inference(
  A,
  pseudo.out,
  debias,
  tau = 1,
  eval.pts = NULL,
  muhat.vals = NULL,
  mhat.obs = NULL,
  ...
)
}
\arguments{
\item{A}{ADD}

\item{pseudo.out}{ADD}

\item{debias}{boolean ADD}

\item{tau}{ratio bandwidths ADD}

\item{eval.pts}{ADD}

\item{muhat.vals}{ADD}

\item{mhat.obs}{ADD}

\item{...}{ADD}
}
\description{
debiased_inference
}
