% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VIMP.R
\name{get_vimp}
\alias{get_vimp}
\title{CATE}
\usage{
get_vimp(
  cate.fit,
  var.names,
  lab.var.names = NULL,
  x = NULL,
  y = NULL,
  a = NULL,
  nsplits = NULL,
  pi.x = NULL,
  mu1.x = NULL,
  mu0.x = NULL,
  drl.x = NULL
)
}
\arguments{
\item{cate.fit}{object from fitting the cate using drl.cate}
}
\description{
This function calculates Variable IMPortance measures following the procedure
in Hines et al (2022) https://arxiv.org/pdf/2204.06030.pdf
}
